[FAGSYSTEM-194276] flytte testdata til egen fil
